third_party/blink/web_tests/platform/generic/editing/pasteboard/paste-when-over-link-expected.txt


This test passes if middle mouse clicking on a link where the event handler
focuses a textarea does not paste into the textarea.


After middle click on link:
|   <shadow:root>
|     <div>
|       "Content"
